requirements.txt
pyscreenshot
pillow
numpy
pynput